requirements.txt
pyserial>=3.5
requests>=2.25.0
websockets>=12.0
aiohttp>=3.9
